SSE2 routine for the GF lookup multiply routine
Not sure if it is actually a lot faster as a whole, but isolated testing showed a ~20% increase in speed. Still a lot of loads, but that is inevitable with table lookups. An option might be to do a PCLMULQDQ routine (SSE4.1), but that will likely be very slow for GF(2^16) compared to lookup tables.