More efficient hash calculation
Calculate the md5 of the first 16kB when opening the input files. The md5 of the entire files and their slices are now calculated while the slices are being read for Reed-Solomon encoding, reducing disk I/O.

Before, the md5 of the entire file was calculated separately from those of the slices, but now they are integrated, resulting in faster computation of the hashes.