Fix in diskfile reading
Behavior does not change when offsets are not used. When offsets are used, we want to set the filesize to the size of the "virtual" file, and therefore file size is a lot smaller than the actual physical file size. The offset in the physical file would then often exceed the file size specified.